bootserver/pyproject.toml
[tool.poetry]
name = "bootserver"
version = "0.1.0"
description = ""
authors = ["'John Syrinek' <john.syrinek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psutil = "^5.9.5"
gitpython = "^3.1.31"
rich = "^13.3.5"
py3tftp = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
